the passport is working, Now have to modularize some areas and work with the routing
