+++ models/user.js
@@ -4,4 +4,6 @@ const user = new mongoose.Schema({
     password:String
 });
 
-module.exports = mongoose.model("User", user);
+const User = mongoose.model("User", user);
+
+module.exports = User;
