+++ server.js
@@ -5,9 +5,9 @@ const cookieParser = require("cookie-parser");
 const bcrypt= require("bcryptjs");
 const session = require("express-session");
 const bodyParser = require("body-parser");
-const User = require("./user");
+const User = require("./models/user");
 
-
+const morgan = require('morgan');
 
 const mongoose = require("mongoose");
 const routes = require("./routes");
@@ -19,11 +19,10 @@ app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
 // Serve up static assets (usually on heroku)
-if (process.env.NODE_ENV === "production") {
-  app.use(express.static("client/build"));
-}
-// Add routes, both API and view
-app.use(routes);
+// if (process.env.NODE_ENV === "production") {
+//   app.use(express.static("client/build"));
+// }
+
 
 // Connect to the Mongo DB
 mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/devconnect", {
@@ -37,11 +36,11 @@ mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/devconnect", {
 // .....................Middleware..................................................
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended:true}));
-
-app.use(cors({
-  origin: "http://localhost:3000",
-  credentials: true
-}));
+app.use(morgan('tiny'));
+// app.use(cors({
+//   origin: "http://localhost:3000",
+//   credentials: true
+// }));
 
 app.use(session({
   secret: "secretcode",
@@ -56,7 +55,7 @@ require("./passportConfig")(passport);
 
 //....................end of Middleware...........................................
 
-app.post ("/", (req, res, next) =>{
+app.post ("/api/", (req, res, next) =>{
   passport.authenticate("local", (err,user,info) =>{
       if (err) throw err;
       if (!user) res.send("no user exists");
@@ -70,7 +69,7 @@ app.post ("/", (req, res, next) =>{
   })(req, res, next);
 });
 
-app.post ("/signup", (req,res) =>{
+app.post ("/api/signup", (req,res) =>{
   User.findOne({username: req.body.username}, async (err,doc) => {
       if (err) throw err;
       if (doc) res.send("User Already Exists");
@@ -87,11 +86,12 @@ app.post ("/signup", (req,res) =>{
   })
 });
 
-app.get ("/message", (req,res) =>{
+app.get ("/api/message", (req,res) =>{
   res.send(req.user);
 })
 
-
+// Add routes, both API and view
+// app.use(routes);
 
 
 
@@ -99,5 +99,5 @@ app.get ("/message", (req,res) =>{
 
 // Start the API server
 app.listen(PORT, function() {
-  console.log(`🌎  ==> API Server now listening on PORT ${PORT}!`);
+  console.log(`🌎  ==> API Server now listening on PORT ${PORT }!`);
 });
